Memecah widget login
